Added log system
Changed VendItem to return a string so that I could use said string create a
and attach choosen input to the log via VendLog.

Then used the System.IO namespace to access the StreamWriter method to create
and print to the log.txt.

Added a fail check for the continue check. Now the method only allows for the correct input.